refactor: use long instead of int on the follow counts

+++ src/main/java/com/gray/bird/user/follow/dto/FollowCounts.java
@@ -4,5 +4,5 @@ import java.util.UUID;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-public record FollowCounts(@JsonIgnore UUID userId, int following, int followers) {
+public record FollowCounts(@JsonIgnore UUID userId, long following, long followers) {
 }
